replies and actions organization
